refactor: restructures components
