+++ app/components/Pages/Feedback/RightSidebar/index.tsx
@@ -1,0 +1,12 @@
+import { Box } from "@chakra-ui/react"
+import React, { FC } from "react"
+
+const RightSidebar: FC = () => {
+  return (
+    <Box w="300px" borderLeftWidth={1} overflowY="hidden">
+      Sidebar
+    </Box>
+  )
+}
+
+export default RightSidebar
