+++ app/components/layouts/index.tsx
@@ -1,22 +1,12 @@
-import { Flex, useColorModeValue } from "@chakra-ui/react"
-import TopNavbar from "app/components/layouts/top-navbar"
+import { Flex } from "@chakra-ui/react"
+import TopNavbar from "app/components/Layouts/TopNavbar"
 import React, { FC } from "react"
 
 const Layout: FC = ({ children }) => {
-  const bgColor = useColorModeValue("brand.100", "brand.900")
-  const color = useColorModeValue("brand.900", "brand.100")
-
   return (
     <>
       <TopNavbar />
-      <Flex
-        h="calc(100vh - 50px)"
-        w="100vw"
-        justifyContent="center"
-        alignItems="center"
-        bg={bgColor}
-        color={color}
-      >
+      <Flex h="calc(100vh - 50px)" w="100vw" justifyContent="center" alignItems="center">
         {children}
       </Flex>
     </>
